+++ shell/command_registry.cpp
@@ -17,6 +17,7 @@
 #include "commands/open.hpp"    // Corrected: was open_command.hpp
 #include "commands/load_plugin_command.hpp"
 #include "commands/unload_plugin_command.hpp"
+#include "commands/exec_command.hpp"
 // Note: If the ChatCommand is also a default command, its header should be included here.
 // #include "commands/chat_command.hpp"
 
@@ -79,6 +80,7 @@ void populate_default_commands(CommandRegistry& registry) {
     registry.register_command(std::make_unique<OpenCommand>());
     registry.register_command(std::make_unique<LoadPluginCommand>(registry));   // Added
     registry.register_command(std::make_unique<UnloadPluginCommand>(registry)); // Added
+    registry.register_command(std::make_unique<ExecCommand>());
     // If ChatCommand is a default command, register it here:
     // registry.register_command(std::make_unique<ChatCommand>());
 }